add interrupt trait
The interrupt trait just has a trigger method. After looking at
current devices used by the consumers of rust-vmm, the only
thing that all devices have in common in respect to interrupts is
the need to trigger an interrupt. Hence, the Interrupt trait is
very simple and only takes care of common devices functionality.

Signed-off-by: Andreea Florescu <fandree@amazon.com>