interrupts: specify mutability of self in traits
Most methods in the traits need to mutate the state. The inital idea
was that save mutability can be achieved by wrapping the trait
implementers in a Mutex. I find this to be misleading when reading
the code. The proposal is to use mut where needed unless there is
a good reason not to.

Signed-off-by: Andreea Florescu <fandree@amazon.com>